refactor(plugins): remove PluginPage
