+++ public/app/features/plugins/utils.ts
@@ -1,0 +1,29 @@
+import { GrafanaPlugin, PanelPluginMeta, PluginType } from '@grafana/data';
+import { getPluginSettings } from './PluginSettingsCache';
+import { importAppPlugin, importDataSourcePlugin } from './plugin_loader';
+import { importPanelPluginFromMeta } from './importPanelPlugin';
+
+export async function loadPlugin(pluginId: string): Promise<GrafanaPlugin> {
+  const info = await getPluginSettings(pluginId);
+  let result: GrafanaPlugin | undefined;
+
+  if (info.type === PluginType.app) {
+    result = await importAppPlugin(info);
+  }
+  if (info.type === PluginType.datasource) {
+    result = await importDataSourcePlugin(info);
+  }
+  if (info.type === PluginType.panel) {
+    const panelPlugin = await importPanelPluginFromMeta(info as PanelPluginMeta);
+    result = (panelPlugin as unknown) as GrafanaPlugin;
+  }
+  if (info.type === PluginType.renderer) {
+    result = { meta: info } as GrafanaPlugin;
+  }
+
+  if (!result) {
+    throw new Error('Unknown Plugin type: ' + info.type);
+  }
+
+  return result;
+}
